+++ R/includeJqueryUI.R
@@ -25,14 +25,5 @@ includeJqueryUI <- function() {
   .Deprecated(
     msg = 'Since v0.2.0, there is no need to call includeJqueryUI() before using other shinyjqui functions'
   )
-
-  shiny::addResourcePath('shinyjqui', system.file('www', package = 'shinyjqui'))
-
-  shiny::singleton(
-    shiny::tags$head(
-      shiny::tags$script(src = "shared/jqueryui/jquery-ui.min.js"),
-      shiny::tags$link(rel = "stylesheet", href = "shared/jqueryui/jquery-ui.css"),
-      shiny::tags$script(src = 'shinyjqui/shinyjqui.min.js')
-    )
-  )
+  jquiHead()
 }